refactor(router):404페이지 이름 수정, router위치 수정
